Remove Providers from stateDataList
