+++ src/RhinoProvider.tsx
@@ -1,13 +1,14 @@
 import * as React from 'react';
 import { createContext, useContext } from 'react';
-import createSingleRhinoState, { ProviderProps } from './create-single-rhino-state';
+import { IProvider, ProviderProps } from './types';
+import createGlobalState from './create-global-state';
+import CombineProviders from './utils/CombineProviders';
 
 interface StateObject {
-  [key: string]: any | (() => any);
+  [key: string]: any;
 }
 
 interface StateData<T> {
-  provider: React.FC<ProviderProps>;
   stateHook: () => T;
   updaterHook: () => React.Dispatch<React.SetStateAction<T>>;
 }
@@ -20,55 +21,31 @@ interface RhinoProviderProps extends ProviderProps {
   initialStates: StateObject;
 }
 
-interface RhinoStateProviderProps extends ProviderProps {
-  stateDataList: StateDataList;
-}
-
-const StateDataContext = createContext<StateObject>({});
-
-/**
- *
- * Provider Component which makes the states available to all the nested components
- */
-const RhinoStateProvider: React.FC<RhinoStateProviderProps> = ({
-  children,
-  stateDataList,
-}: RhinoStateProviderProps) => {
-  const ProviderList = Object.values(stateDataList).map((stateData) => stateData.provider);
-
-  // Combining all providers into a single provider
-  const RootProvider = ProviderList.reduce((ProviderAccumulator, Provider) => {
-    return ({ children }: ProviderProps) => (
-      <ProviderAccumulator>
-        <Provider>{children}</Provider>
-      </ProviderAccumulator>
-    );
-  });
-
-  return <RootProvider>{children}</RootProvider>;
-};
+const StateDataContext = createContext<StateDataList>({});
 
 const RhinoProvider: React.FC<RhinoProviderProps> = ({
   children,
   initialStates: initialStatesObject,
 }: RhinoProviderProps) => {
   const stateDataList: StateDataList = {};
+  const ProviderList: Array<IProvider> = [];
 
   Object.keys(initialStatesObject).forEach((key) => {
     const initialValue = initialStatesObject[key];
-    const { Provider, useStateValue, useStateUpdate } = createSingleRhinoState(initialValue);
+    const { Provider, useStateValue, useStateUpdate } = createGlobalState(initialValue);
 
     stateDataList[key] = {
-      provider: Provider,
       stateHook: useStateValue,
       updaterHook: useStateUpdate,
     };
+
+    ProviderList.push(Provider);
   });
 
   return (
-    <StateDataContext.Provider value={stateDataList}>
-      <RhinoStateProvider stateDataList={stateDataList}>{children}</RhinoStateProvider>
-    </StateDataContext.Provider>
+    <CombineProviders providers={ProviderList}>
+      <StateDataContext.Provider value={stateDataList}>{children}</StateDataContext.Provider>
+    </CombineProviders>
   );
 };
