+++ src/types.ts
@@ -1,0 +1,5 @@
+export interface ProviderProps {
+  children: React.ReactChild | React.ReactChild[];
+}
+
+export type IProvider = React.FC<ProviderProps>;
